fix add row and load or whatever

+++ VidInjector9002.5/src/formatting.cpp
@@ -41,7 +41,9 @@ std::string fixDoubleQuote(std::string str) {
 }
 
 std::error_code copyfile(std::string inpath, std::string outpath) {//also works with directories
-	std::error_code error;
+	std::error_code error(0, std::system_category());
+	if(strcmp(inpath.c_str(), outpath.c_str()) == 0)
+		return error;
 	std::filesystem::remove_all(std::filesystem::path((const char8_t*)&*outpath.c_str()), error);
 	if (error) return error;
 	std::filesystem::copy(std::filesystem::path((const char8_t*)&*inpath.c_str()), std::filesystem::path((const char8_t*)&*outpath.c_str()), std::filesystem::copy_options::recursive, error);
